Self-driving-Car-Using-virtual carRaspberry-Pi - Copy/requirements.txt
# Core Dependencies
opencv-python>=4.5.0
numpy>=1.19.0
ultralytics>=8.0.0

# Raspberry Pi GPIO (optional, for hardware)
RPi.GPIO>=0.7.0

# Environment variables
python-dotenv>=0.19.0

# Optional: For advanced features
scipy>=1.5.0
scikit-image>=0.17.0
Pillow>=8.0.0

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.10.0
black>=21.0
flake8>=3.9.0

